refactor: extract XmlMapper into separate class
