+++ app/src/main/java/com/example/appaddressbook/di/AppModule.kt
@@ -2,6 +2,7 @@ package com.example.appaddressbook.di
 
 import com.example.appaddressbook.contacts_loader.ContactsExporter
 import com.example.appaddressbook.contacts_loader.ContactsLoader
+import com.example.appaddressbook.contacts_loader.AddressBookXmlMapper
 import com.example.appaddressbook.repository.ContactsRepository
 import com.example.appaddressbook.repository.ContactsRepositoryImpl
 import dagger.Module
@@ -22,13 +23,19 @@ object AppModule {
 
     @Provides
     @Singleton
-    fun provideContactLoader(): ContactsLoader {
-        return ContactsLoader()
+    fun provideXmlMapper(): AddressBookXmlMapper {
+        return AddressBookXmlMapper()
     }
 
     @Provides
     @Singleton
-    fun provideContactExporter(): ContactsExporter {
-        return ContactsExporter()
+    fun provideContactLoader(xmlMapper: AddressBookXmlMapper): ContactsLoader {
+        return ContactsLoader(xmlMapper)
+    }
+
+    @Provides
+    @Singleton
+    fun provideContactExporter(xmlMapper: AddressBookXmlMapper): ContactsExporter {
+        return ContactsExporter(xmlMapper)
     }
 }
